requirements.txt
pandas==2.2.2
numpy==2.0.1
matplotlib==3.9.0
seaborn==0.13.2
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
notebook==7.2.0

